+++ source/marketparticipant/Energinet.DataHub.MarketParticipant.Application/Handlers/UserRoles/DeactivateUserRoleHandler.cs
@@ -16,7 +16,10 @@ using System;
 using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
+using Energinet.DataHub.Core.App.Common.Abstractions.Users;
 using Energinet.DataHub.MarketParticipant.Application.Commands.UserRoles;
+using Energinet.DataHub.MarketParticipant.Application.Security;
+using Energinet.DataHub.MarketParticipant.Application.Services;
 using Energinet.DataHub.MarketParticipant.Domain;
 using Energinet.DataHub.MarketParticipant.Domain.Exception;
 using Energinet.DataHub.MarketParticipant.Domain.Model.Users;
@@ -28,17 +31,29 @@ namespace Energinet.DataHub.MarketParticipant.Application.Handlers.UserRoles;
 public sealed class DeactivateUserRoleHandler : IRequestHandler<DeactivateUserRoleCommand>
 {
     private readonly IUserRepository _userRepository;
+    private readonly IUserRoleAssignmentAuditLogEntryRepository _userRoleAssignmentAuditLogEntryRepository;
     private readonly IUserRoleRepository _userRoleRepository;
     private readonly IUnitOfWorkProvider _unitOfWorkProvider;
+    private readonly IUserRoleAuditLogService _userRoleAuditLogService;
+    private readonly IUserRoleAuditLogEntryRepository _userRoleAuditLogEntryRepository;
+    private readonly IUserContext<FrontendUser> _userContext;
 
     public DeactivateUserRoleHandler(
         IUserRepository userRepository,
+        IUserRoleAssignmentAuditLogEntryRepository userRoleAssignmentAuditLogEntryRepository,
+        IUserContext<FrontendUser> userContext,
         IUserRoleRepository userRoleRepository,
-        IUnitOfWorkProvider unitOfWorkProvider)
+        IUnitOfWorkProvider unitOfWorkProvider,
+        IUserRoleAuditLogService userRoleAuditLogService,
+        IUserRoleAuditLogEntryRepository userRoleAuditLogEntryRepository)
     {
         _userRepository = userRepository;
+        _userRoleAssignmentAuditLogEntryRepository = userRoleAssignmentAuditLogEntryRepository;
+        _userContext = userContext;
         _userRoleRepository = userRoleRepository;
         _unitOfWorkProvider = unitOfWorkProvider;
+        _userRoleAuditLogService = userRoleAuditLogService;
+        _userRoleAuditLogEntryRepository = userRoleAuditLogEntryRepository;
     }
 
     public async Task Handle(DeactivateUserRoleCommand request, CancellationToken cancellationToken)
@@ -64,13 +79,47 @@ public sealed class DeactivateUserRoleHandler : IRequestHandler<DeactivateUserRo
                 var role = user.RoleAssignments.Single(x => x.UserRoleId == userRoleId);
                 user.RoleAssignments.Remove(role);
 
+                await AuditRoleAssignmentAsync(user, role).ConfigureAwait(false);
                 await _userRepository.AddOrUpdateAsync(user).ConfigureAwait(false);
             }
+
+            var userRoleInitStateForAuditLog = CopyUserRoleForAuditLog(userRole);
 
             userRole.Status = UserRoleStatus.Inactive;
             await _userRoleRepository.UpdateAsync(userRole).ConfigureAwait(false);
 
+            var auditLogs = _userRoleAuditLogService.BuildAuditLogsForUserRoleChanged(
+                new UserId(request.ChangedByUserId),
+                userRoleInitStateForAuditLog,
+                userRole);
+
+            await _userRoleAuditLogEntryRepository.InsertAuditLogEntriesAsync(auditLogs).ConfigureAwait(false);
+
             await uow.CommitAsync().ConfigureAwait(false);
         }
+    }
+
+    private static UserRole CopyUserRoleForAuditLog(UserRole userRole)
+    {
+        return new UserRole(
+            userRole.Id,
+            userRole.Name,
+            userRole.Description,
+            userRole.Status,
+            userRole.Permissions,
+            userRole.EicFunction);
+    }
+
+    private async Task AuditRoleAssignmentAsync(Domain.Model.Users.User user, UserRoleAssignment userRoleAssignment)
+    {
+        await _userRoleAssignmentAuditLogEntryRepository.InsertAuditLogEntryAsync(
+            user.Id,
+            new UserRoleAssignmentAuditLogEntry(
+                user.Id,
+                userRoleAssignment.ActorId,
+                userRoleAssignment.UserRoleId,
+                new AuditIdentity(_userContext.CurrentUser.UserId),
+                DateTimeOffset.UtcNow,
+                UserRoleAssignmentTypeAuditLog.RemovedDueToDeactivation)).ConfigureAwait(false);
     }
 }