+++ source/api-client/Energinet.DataHub.MarketParticipant.Client/documents/release-notes/release-notes.md
@@ -1,5 +1,13 @@
 # Energinet.DataHub.MarketParticipant.Client Release notes
 
+## Version 2.32.1
+
+- Added /audit-identity controller.
+
+## Version 2.32.0
+
+- Changed ChangedByUserId to AuditIdentityId for audited types.
+
 ## Version 2.31.8
 
 - Bumped .github to use v12